Version 0.17.0 - Refactored the portfolio/show page: updated some CSS classes and added a JavaScript script that moves portfolio content depending on screen size. - Created reusable components project-single and service-single for use both in the public part and in admin preview. - Implemented a single project detail view. - Implemented the projects index view with pagination. - Created a custom pagination template for the public section. - Updated the sorting logic for projects and services: items with position = NULL now appear after items with a defined position. - Created reusable components services-section and portfolio-section for use on the homepage, index pages, and preview pages. Added props to support different use cases. - Replaced the default browser drag icon in the input-gallery.js drag-and-drop component. - Adjusted modal blade component styles to ensure the modal always appears above all other DOM elements. - Implemented portfolio preview (single and grid) views in the admin. - Refactored service preview (single and grid) views in the admin. - Added redirection to the index page if an item is not found or unpublished in the public part, and abort(404) in the admin if the item is not found.

+++ app/Http/Controllers/Admin/PortfolioController.php
@@ -59,9 +59,22 @@ class PortfolioController extends Controller
     /**
      * Display the specified resource.
      */
-    public function show(Portfolio $portfolio)
+    public function showSingle(Portfolio $portfolio)
     {
-        //
+        return view('admin.portfolio.preview-single', ['portfolio' => $portfolio]);
+
+    }
+
+    public function showGrid(Portfolio $portfolio){
+
+        $portfolios = [];
+
+        for($i = 0; $i < 4; $i++){
+            $portfolios[] = $portfolio;
+        }
+
+        return view('admin.portfolio.preview-grid', ['portfolios' => $portfolios]);
+
     }
 
     /**
@@ -108,6 +121,9 @@ class PortfolioController extends Controller
         $portfolio->updateImages($positions, $images, $oldImagesIds, $existingImages);
         $portfolio->deleteImages($existingImages, $oldImagesIds);
 
+        //Project update
+        $portfolio->update($validated);
+
         Cache::forget('portfolios');
 
         return redirect(route('admin.portfolio.index'))
@@ -125,7 +141,10 @@ class PortfolioController extends Controller
 
         $portfolio = Portfolio::findOrFail($id);
 
-        Storage::delete($portfolio->thumbnail);
+        if($portfolio->thumbnail) {
+            Storage::delete($portfolio->thumbnail);
+        }
+
 
         $existingImages = $portfolio->images()->get();
